+++ src/append/journald/mod.rs
@@ -294,7 +294,7 @@ impl Append for Journald {
         );
         // Put all structured values of the record
         let mut visitor = WriteKeyValues(&mut buffer);
-        record.visit_kvs(&mut visitor)?;
+        record.key_values().visit(&mut visitor)?;
         for d in diags {
             d.visit(&mut visitor)?;
         }