+++ src/kv.rs
@@ -14,10 +14,12 @@
 
 //! The module for key-value pairs in a log record or a diagnostic context.
 
+// This file is derived from https://github.com/SpriteOvO/spdlog-rs/blob/788bda33/spdlog/src/kv.rs
+
 pub extern crate value_bag;
 
 use std::fmt;
-use std::fmt::Debug;
+use std::slice;
 
 use value_bag::OwnedValueBag;
 use value_bag::ValueBag;
@@ -95,5 +97,115 @@ impl KeyOwned {
 impl fmt::Display for KeyOwned {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Display::fmt(&self.0, f)
+    }
+}
+
+/// A collection of key-value pairs.
+pub struct KeyValues<'a>(KeyValuesState<'a>);
+
+enum KeyValuesState<'a> {
+    Borrowed(&'a [(Key<'a>, Value<'a>)]),
+    Owned(&'a [(KeyOwned, ValueOwned)]),
+}
+
+impl<'a> KeyValues<'a> {
+    /// Get the number of key-value pairs.
+    pub fn len(&self) -> usize {
+        match self.0 {
+            KeyValuesState::Borrowed(p) => p.len(),
+            KeyValuesState::Owned(p) => p.len(),
+        }
+    }
+
+    /// Check if there are no key-value pairs.
+    pub fn is_empty(&self) -> bool {
+        match self.0 {
+            KeyValuesState::Borrowed(p) => p.is_empty(),
+            KeyValuesState::Owned(p) => p.is_empty(),
+        }
+    }
+
+    /// Get an iterator over the key-value pairs.
+    pub fn iter(&self) -> KeyValuesIter<'a> {
+        match &self.0 {
+            KeyValuesState::Borrowed(p) => KeyValuesIter(KeyValuesIterState::Borrowed(p.iter())),
+            KeyValuesState::Owned(p) => KeyValuesIter(KeyValuesIterState::Owned(p.iter())),
+        }
+    }
+
+    /// Visit the key-value pairs with the provided visitor.
+    pub fn visit(&self, visitor: &mut dyn Visitor) -> Result<(), Error> {
+        for (k, v) in self.iter() {
+            visitor.visit(k, v)?;
+        }
+        Ok(())
+    }
+}
+
+impl fmt::Debug for KeyValues<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_list().entries(self.iter()).finish()
+    }
+}
+
+impl Clone for KeyValues<'_> {
+    fn clone(&self) -> Self {
+        match &self.0 {
+            KeyValuesState::Borrowed(p) => KeyValues(KeyValuesState::Borrowed(p)),
+            KeyValuesState::Owned(p) => KeyValues(KeyValuesState::Owned(p)),
+        }
+    }
+}
+
+impl Default for KeyValues<'_> {
+    fn default() -> Self {
+        KeyValues(KeyValuesState::Borrowed(&[]))
+    }
+}
+
+impl<'a> IntoIterator for KeyValues<'a> {
+    type Item = (Key<'a>, Value<'a>);
+    type IntoIter = KeyValuesIter<'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.iter()
+    }
+}
+
+impl<'a> From<&'a [(Key<'a>, Value<'a>)]> for KeyValues<'a> {
+    fn from(kvs: &'a [(Key<'a>, Value<'a>)]) -> Self {
+        Self(KeyValuesState::Borrowed(kvs))
+    }
+}
+
+impl<'a> From<&'a [(KeyOwned, ValueOwned)]> for KeyValues<'a> {
+    fn from(kvs: &'a [(KeyOwned, ValueOwned)]) -> Self {
+        Self(KeyValuesState::Owned(kvs))
+    }
+}
+
+/// An iterator over key-value pairs.
+pub struct KeyValuesIter<'a>(KeyValuesIterState<'a>);
+
+enum KeyValuesIterState<'a> {
+    Borrowed(slice::Iter<'a, (Key<'a>, Value<'a>)>),
+    Owned(slice::Iter<'a, (KeyOwned, ValueOwned)>),
+}
+
+impl<'a> Iterator for KeyValuesIter<'a> {
+    type Item = (Key<'a>, Value<'a>);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        match &mut self.0 {
+            KeyValuesIterState::Borrowed(iter) => iter.next().map(|(k, v)| (k.clone(), v.clone())),
+            KeyValuesIterState::Owned(iter) => iter.next().map(|(k, v)| (k.by_ref(), v.by_ref())),
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        match &self.0 {
+            KeyValuesIterState::Borrowed(iter) => iter.size_hint(),
+            KeyValuesIterState::Owned(iter) => iter.size_hint(),
+        }
     }
 }