+++ src/layout/google_cloud_logging.rs
@@ -221,7 +221,7 @@ impl Layout for GoogleCloudLoggingLayout {
             trace_sampled: None,
         };
 
-        record.visit_kvs(&mut visitor)?;
+        record.key_values().visit(&mut visitor)?;
         for d in diags {
             d.visit(&mut visitor)?;
         }